Replace get-accounts with get-accounts-v2
