+++ .github/workflows/check.yml
@@ -60,9 +60,8 @@ jobs:
           python-version: '3.10' 
       - run: pip install -r requirements.txt
 
-      - run: ./get-accounts.py
-      - run: ./get-accounts-v2-1.py 'non-production'
-      - run: cat accounts-v2.csv
+      - run: ./get-accounts.py 'non-production'
+      - run: cat accounts.csv
 
       - name: Set up AWS credentials
         uses: aws-actions/configure-aws-credentials@v4
@@ -111,9 +110,8 @@ jobs:
           python-version: '3.10' 
       - run: pip install -r requirements.txt
 
-      - run: ./get-accounts.py
-      - run: ./get-accounts-v2-1.py 'production'
-      - run: cat accounts-v2.csv
+      - run: ./get-accounts.py 'production'
+      - run: cat accounts.csv
 
       - name: Set up AWS credentials
         uses: aws-actions/configure-aws-credentials@v4
