+++ .github/workflows/deploy.yml
@@ -33,7 +33,7 @@ jobs:
       - name: Set up AWS credentials
         uses: aws-actions/configure-aws-credentials@v4
         with:
-          role-to-assume: ${{ secrets.GHA_ROLE_ARN }}
+          role-to-assume: arn:aws:iam::892537467220:role/ObservationConfigurationGitHubIde-GitHubActionsRole-MdPyTyImAExz
           aws-region: eu-west-2
 
       - uses: actions/setup-python@v5
@@ -41,7 +41,13 @@ jobs:
           python-version: '3.10' 
       - run: pip install -r requirements.txt
 
-      - run: ./get-accounts.py
+      - run: ./get-accounts.py 'non-production'
+
+      - name: Set up AWS credentials
+        uses: aws-actions/configure-aws-credentials@v4
+        with:
+          role-to-assume: ${{ secrets.GHA_ROLE_ARN }}
+          aws-region: eu-west-2
 
       - uses: hashicorp/setup-terraform@v3
         with:
@@ -76,7 +82,7 @@ jobs:
       - name: Set up AWS credentials
         uses: aws-actions/configure-aws-credentials@v4
         with:
-          role-to-assume: ${{ secrets.GHA_ROLE_ARN }}
+          role-to-assume: arn:aws:iam::892537467220:role/ObservationConfigurationGitHubIde-GitHubActionsRole-MdPyTyImAExz
           aws-region: eu-west-2
 
       - uses: actions/setup-python@v5
@@ -84,7 +90,13 @@ jobs:
           python-version: '3.10' 
       - run: pip install -r requirements.txt
 
-      - run: ./get-accounts.py
+      - run: ./get-accounts.py 'production'
+
+      - name: Set up AWS credentials
+        uses: aws-actions/configure-aws-credentials@v4
+        with:
+          role-to-assume: ${{ secrets.GHA_ROLE_ARN }}
+          aws-region: eu-west-2
 
       - uses: hashicorp/setup-terraform@v3
         with:
